fix: Follow review
- Move lightline.colorscheme.* to nightfox.util.lightline
- Support vim
- Correct readme
