+++ lua/nightfox/util/lightline.lua
@@ -1,0 +1,87 @@
+local M = {}
+
+local Color = require("nightfox.lib.color")
+
+---@param name string
+---@return table
+function M.generate(name)
+  local spec = require("nightfox.spec").load(name)
+  local pal = spec.palette
+  local bg = Color.from_hex(spec.bg0)
+  local fg = spec.fg2
+
+  ---@param color string
+  ---@param amount number|nil
+  ---@return table
+  local function generate_mode(color, amount)
+    amount = amount or 0.3
+    local fade = bg:blend(Color.from_hex(color), amount):to_css()
+    local b = bg:to_css()
+
+    return {
+      left = { { b, color }, { fg, fade } },
+      right = { { fg, fade } },
+      middle = { { fg, b } },
+    }
+  end
+
+  -- stylua: ignore
+  return {
+    normal   = generate_mode(pal.blue.base),
+    insert   = generate_mode(pal.green.base),
+    command  = generate_mode(pal.yellow.base),
+    visual   = generate_mode(pal.magenta.base),
+    replace  = generate_mode(pal.red.base),
+    inactive = generate_mode(spec.fg3),
+    tabline  = generate_mode(spec.fg3),
+  }
+end
+
+---@param name string
+---@return string
+function M.dump(name)
+  local obj = M.generate(name)
+  if type(obj) ~= 'table' then
+    return "'" .. tostring(obj) .. "'"
+  end
+
+  local i = 0
+  local ret = ''
+  local is_arr = 0
+  for k,v in pairs(obj) do
+    if type(k) ~= 'number' then
+      -- the `obj` is a dict
+      if i == 0 then
+        -- first item
+        ret = '{'
+      else
+        ret = ret .. ', '
+      end
+
+      -- wrap the key with single quotes
+      k = "'" .. k .. "'"
+      ret = ret .. k .. ': ' .. M.dump(v)
+    else
+      -- the `obj` is an array
+      if i == 0 then
+        -- first item
+        ret = '['
+        is_arr = 1
+      else
+        ret = ret .. ', '
+      end
+
+      ret = ret .. M.dump(v)
+    end
+
+    i = i + 1
+  end
+
+  if is_arr == 0 then
+    return ret .. '}'
+  else
+    return ret .. ']'
+  end
+end
+
+return M
